Cleanup FixMessageParser#parse()
Use the fact that Connection#parse() sets buffer mark at the beginning of the
next message and use buffer.reset() in FixMessageParser#parse() when we've made
sure there's a full message waiting for us. This reduces the need to rebuild a
byte buffer of the message after we've figured out the message size.

Signed-off-by: Pekka Enberg <penberg@reaktor.fi>