silvertip: Add FIX message
A FIX message comprises of the whole message data as well as the type of
the message, MsgType(35). Also, MsgSeqNum(34) and the time of when the
message has been received can be assigned to FixMessage.

MsgType(35) is needed for determining whether the message should be
placed in the message queue through which any gaps in the sequence of
messages is detected using the MsgSeqNum(34) field.

Declare a dependency to Joda-Time.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>