silvertip: Check validity of keys before using them
The following is stated in the API documentation of SelectionKey:

 Cancelling a key does not immediately remove it from its selector; it is
 instead added to the selector's cancelled-key set for removal during the next
 selection operation. The validity of a key may be tested by invoking its
 isValid method.

Thus, the underlying channel may be closed, yet the keys returned by
Selector#keys may still contain the key associated with the closed
channel, resulting to a CancelledKeyException upon accessing it. In
order to make sure that the key is not cancelled, SelectionKey#isValid
must be invoked prior to other invocations.

[ karim.osman@reaktor.fi: Revise commit message ]
Signed-off-by: Johannes Laire <johannes.laire@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>