silvertip: Timeout in 'Events#processNow'
When using exclusively the non-blocking 'Events#processNow' method,
the 'EventSource#timeout' callback is never called. Fix this by
calling the callback every time the selected key set is empty. This
is not optimal: in many applications 'Events#processNow' is invoked
in rapid succession, and in most of these invocations the selected
key set is empty.

Signed-off-by: Jussi Virtanen <jussi.virtanen@valotrading.com>